feat(posts): moved homepage post list into its own component
